% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Bushfire Data Explorer App}
\usage{
run_app()
}
\value{
A Shiny app object
}
\description{
Launch the Bushfire Data Explorer App
}
\examples{
if (interactive()) {
  run_app()
}
}
